refactor: rename HomeScreenCarousel to Carousel and use it on search results
The carousel isn't home-specific anymore — search results for albums
and artists now show the same horizontal scrollable card layout instead
of a two-column grid.

- Move components/screens/home/HomeScreenCarousel.vue (+.spec.ts) to
  components/ui/Carousel.vue (+.spec.ts) and drop the HomeScreen prefix.
- Update the six home carousels (New/Top/Random × Albums/Artists) to
  import the new path and tag name.
- Replace the <ul class="results"> grid in AlbumExcerptResultsBlock and
  ArtistExcerptResultsBlock with the Carousel, matching the home-screen
  look and reusing the same chevron + (eventually) refresh affordances.
  ExcerptResultBlock is no longer needed at those two call sites.
