+++ resources/assets/js/components/screens/home/RandomArtists.vue
@@ -1,5 +1,5 @@
 <template>
-  <HomeScreenCarousel :on-refresh="refresh">
+  <Carousel :on-refresh="refresh">
     <template #header>Random Artists</template>
     <template v-if="loading">
       <ArtistCardSkeleton v-for="i in 6" :key="i" />
@@ -8,7 +8,7 @@
       <ArtistCard v-for="artist in artists" :key="artist.id" :artist />
     </template>
     <p v-else class="text-k-fg-50">No artists yet.</p>
-  </HomeScreenCarousel>
+  </Carousel>
 </template>
 
 <script lang="ts" setup>
@@ -17,7 +17,7 @@ import { overviewStore } from '@/stores/overviewStore'
 
 import ArtistCard from '@/components/artist/ArtistCard.vue'
 import ArtistCardSkeleton from '@/components/ui/album-artist/ArtistAlbumCardSkeleton.vue'
-import HomeScreenCarousel from '@/components/screens/home/HomeScreenCarousel.vue'
+import Carousel from '@/components/ui/Carousel.vue'
 
 const props = withDefaults(defineProps<{ loading?: boolean }>(), { loading: false })
 const { loading } = toRefs(props)
