+++ resources/assets/js/components/screens/search/AlbumExcerptResultsBlock.vue
@@ -1,27 +1,21 @@
 <template>
-  <ExcerptResultBlock>
+  <Carousel>
     <template #header>Albums</template>
 
-    <ul v-if="searching" class="results">
-      <li v-for="i in 6" :key="i">
-        <AlbumCardSkeleton />
-      </li>
-    </ul>
-    <template v-else>
-      <ul v-if="albums.length" class="results">
-        <li v-for="album in albums" :key="album.id">
-          <AlbumCard :album />
-        </li>
-      </ul>
-      <p v-else>None found.</p>
+    <template v-if="searching">
+      <AlbumCardSkeleton v-for="i in 6" :key="i" />
     </template>
-  </ExcerptResultBlock>
+    <template v-else-if="albums.length">
+      <AlbumCard v-for="album in albums" :key="album.id" :album />
+    </template>
+    <p v-else class="text-k-fg-50">None found.</p>
+  </Carousel>
 </template>
 
 <script lang="ts" setup>
 import { toRefs } from 'vue'
 
-import ExcerptResultBlock from '@/components/screens/search/ExcerptResultBlock.vue'
+import Carousel from '@/components/ui/Carousel.vue'
 import AlbumCardSkeleton from '@/components/ui/album-artist/ArtistAlbumCardSkeleton.vue'
 import AlbumCard from '@/components/album/AlbumCard.vue'
 
@@ -32,10 +26,3 @@ const props = withDefaults(defineProps<{ albums?: Album[]; searching?: boolean }
 
 const { albums, searching } = toRefs(props)
 </script>
-
-<style lang="postcss" scoped>
-@reference '@css/app.pcss';
-.results {
-  @apply grid grid-cols-1 md:grid-cols-2 gap-x-4 gap-y-3;
-}
-</style>
