+++ resources/assets/js/components/screens/search/ArtistExcerptResultsBlock.vue
@@ -1,21 +1,15 @@
 <template>
-  <ExcerptResultBlock>
+  <Carousel>
     <template #header>Artists</template>
 
-    <ul v-if="searching" class="results">
-      <li v-for="i in 6" :key="i">
-        <ArtistCardSkeleton />
-      </li>
-    </ul>
-    <template v-else>
-      <ul v-if="artists.length" class="results">
-        <li v-for="artist in artists" :key="artist.id">
-          <ArtistCard :artist />
-        </li>
-      </ul>
-      <p v-else>None found.</p>
+    <template v-if="searching">
+      <ArtistCardSkeleton v-for="i in 6" :key="i" />
     </template>
-  </ExcerptResultBlock>
+    <template v-else-if="artists.length">
+      <ArtistCard v-for="artist in artists" :key="artist.id" :artist />
+    </template>
+    <p v-else class="text-k-fg-50">None found.</p>
+  </Carousel>
 </template>
 
 <script lang="ts" setup>
@@ -23,7 +17,7 @@ import { toRefs } from 'vue'
 
 import ArtistCard from '@/components/artist/ArtistCard.vue'
 import ArtistCardSkeleton from '@/components/ui/album-artist/ArtistAlbumCardSkeleton.vue'
-import ExcerptResultBlock from '@/components/screens/search/ExcerptResultBlock.vue'
+import Carousel from '@/components/ui/Carousel.vue'
 
 const props = withDefaults(defineProps<{ artists?: Artist[]; searching?: boolean }>(), {
   artists: () => [],
@@ -32,10 +26,3 @@ const props = withDefaults(defineProps<{ artists?: Artist[]; searching?: boolean
 
 const { artists, searching } = toRefs(props)
 </script>
-
-<style lang="postcss" scoped>
-@reference '@css/app.pcss';
-.results {
-  @apply grid grid-cols-1 md:grid-cols-2 gap-x-4 gap-y-3;
-}
-</style>
